Reorganized Warehouse block test cases + new tests for DTO factories
