+++ tests/Warehouse/StockDataFactoryTest.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Grixu\SociusModels\Tests\Warehouse;
+
+use Grixu\SociusModels\Tests\Helpers\DataFactoryTestCase;
+use Grixu\SociusModels\Warehouse\Factories\StockDataFactory;
+
+class StockDataFactoryTest extends DataFactoryTestCase
+{
+    protected string $factory = StockDataFactory::class;
+}
