+++ tests/Warehouse/WarehouseDataFactoryTest.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Grixu\SociusModels\Tests\Warehouse;
+
+use Grixu\SociusModels\Tests\Helpers\DataFactoryTestCase;
+use Grixu\SociusModels\Warehouse\Factories\WarehouseDataFactory;
+
+class WarehouseDataFactoryTest extends DataFactoryTestCase
+{
+    protected string $factory = WarehouseDataFactory::class;
+}
