+++ tests/Warehouse/WarehouseMigrationTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Grixu\SociusModels\Tests;
+namespace Grixu\SociusModels\Tests\Warehouse;
 
 use CreateCustomersTable;
 use CreateWarehouseTable;
@@ -15,14 +15,14 @@ class WarehouseMigrationTest extends MigrationTest
     {
         parent::setUp();
 
-        require_once __DIR__ . '/../migrations/warehouse/2020_10_02_063622_create_warehouse_table.php';
+        require_once __DIR__ . '/../../migrations/warehouse/2020_10_02_063622_create_warehouse_table.php';
         $this->obj = new CreateWarehouseTable();
     }
 
     /** @test */
     public function customer_relationship(): void
     {
-        require_once __DIR__ . '/../migrations/customer/2020_09_30_102037_create_customers_table.php';
+        require_once __DIR__ . '/../../migrations/customer/2020_09_30_102037_create_customers_table.php';
         (new CreateCustomersTable())->up();
 
         $this->up();
